+++ src/ideabox/policy/content/campaign.py
@@ -12,6 +12,14 @@ class ICampaign(model.Schema):
         title=_(u"Enable / Disable project submission"), default=False
     )
 
+    emails = schema.TextLine(
+        title=_(u"Email addresses"),
+        description=_(
+            u'Used to send notification when a new project is proposed by an user. Accept many email addresses separated with ";"'
+        ),
+        required=False,
+    )
+
 
 @implementer(ICampaign)
 class Campaign(Container):